app/requirements.txt
boto3==1.34.64
streamlit==1.32.0
langchain==0.1.12
langchain-community
faiss-cpu==1.8.0
awscli
pypdf
python-dotenv